+++ binary-parser-bench/HttpReq.hs
@@ -13,7 +13,6 @@ import qualified Data.Attoparsec.ByteString as AP
 import qualified Data.Attoparsec.ByteString.Char8 as APC
 import qualified Data.ByteString.Char8 as BC
 import qualified Data.Binary.Parser as BP
-import qualified Data.Binary.Parser.Char8 as BPC
 import Network.HTTP.Types.Version (HttpVersion, http11)
 import qualified Scanner as SC
 
@@ -59,20 +58,21 @@ attoManyHeader = do
 --------------------------------------------------------------------------------
 
 bpHeader = do
-  name <- BPC.takeWhile1 isHeaderChar <* BPC.char ':' <* BP.skipSpaces
+  name <- BP.takeWhile1 isHeaderWord8 <* BP.word8 (c2w ':') <* BP.skipSpaces
   body <- bpBodyLine
   return (name, body)
   where
-    isHeaderChar c = ('a' <= c && c <= 'z')
-                  || ('A' <= c && c <= 'Z')
-                  || ('0' <= c && c <= '0')
-                  || c == '_' || c == '-'
+    isHeaderWord8 w = let c = w2c w in
+                   ('a' <= c && c <= 'z')
+                || ('A' <= c && c <= 'Z')
+                || ('0' <= c && c <= '0')
+                || c == '_' || c == '-'
 
-bpBodyLine = BPC.takeTill (\c -> c == '\r' || c == '\n') <* BP.endOfLine
+bpBodyLine = BP.takeTill (\w -> w == 13 || w == 10) <* BP.endOfLine
 
 bpReqLine = do
-  m <- (BPC.takeTill BPC.isSpace <* BPC.char ' ')
-  (p,q) <- BC.break (=='?') <$> (BPC.takeTill BPC.isSpace <* BPC.char ' ')
+  m <- (BP.takeTill BP.isSpace <* BP.word8 (c2w ' '))
+  (p,q) <- BC.break (=='?') <$> (BP.takeTill BP.isSpace <* BP.word8 (c2w ' '))
   v <- bpHttpVersion
   return (m,p,q,v)
 
@@ -81,7 +81,8 @@ bpHttpVersion = http11 <$ BP.string "HTTP/1.1"
 bpRequest = (,) <$> (bpReqLine <* BP.endOfLine) <*> bpManyHeader
 
 bpManyHeader = do
-  c <- BPC.peek
+  w <- BP.peek
+  let c = w2c w
   if c == '\r' || c == '\n'
     then return []
     else (:) <$> bpHeader <*> bpManyHeader